requirements.txt
opencv-python>=4.9.0
Pillow>=10.2.0
tqdm>=4.66.0
